Lots of new files that are used for creating limits with the incorporation of new parameterizations for various error parameters
ra2bRoostatsClass7_noNSig and ra2bRoostatsClass7_noNSig_withKristen are the definitions of the likelihood files.  They also have a likelihood with no dependence on the SIG measurement to calculate the expected number of results and its variance.

Beta distributions are used for parameters bounded between 0 and 1

Gamma distributions are used for parameters bounded at 0

Correlated errors are modelled as one gamma distributed parameter with mode 1 that undergoes a power law transform to change widths

Also included are various scripts and file structures for running over signal models at CERN on the batch system

the command ./submit_jobs T1bbbb will scan over the T1bbbb model and submit a job for every point in reducedModel.T1bbbb.txt

editing the file will allow for scanning over any signal model.

The results are stored in the results folder.

This hasn't been tested in this directory so YMMV (there may be residual hard-coding in the scripts that I have yet to remove)

NB: RooGammaPdf is a different implementation from the default (version 5.33) ROOT version that does number generation correctly

NB: DebuggingToyMCSampler and HybridToyMCSampler work with ROOT version 5.33 but are not guaranteed to work with later versions (and don't work with some earlier versions)

NB: ProfileLikelihoodTestStat_New is very similar to ProfileLikelihoodTestStat but has some extra checking built in to make sure that its value is never negative