Owen : Major revision of toymc2b.c  Most things are stored in 3D arrays.   Here are some Draw examples of using the arrays:
Fit value of the total number of SUSY 0lep events (including eff and btag corrections).

  toytt->Draw("Sum$(fit_susy_0lep_3da)")

Fit value of the total number of SUSY 0lep >=3b events (including eff and btag corrections).

  toytt->Draw("Sum$(fit_susy_0lep_3da[][][2])")

Fit value of the total number of SUSY 0lep >=3b events in the highest MET bin,
if there are 4 MET bins (including eff and btag corrections).

  toytt->Draw("Sum$(fit_susy_0lep_3da[3][][2])")

True value of the total number of SUSY 0lep >=3b events in the highest MET bin,
if there are 4 MET bins.

  toytt->Draw("Sum$(mcval_susy_0lep_3da[3][][2])")

Fit value of the total number of SUSY 0lep >=3b events in the highest HT bin,
if there are 4 HT bins (including eff and btag corrections).

  toytt->Draw("Sum$(fit_susy_0lep_3da[][3][2])")


Fit value of the total number of SUSY 0lep >=3b events in
the highest two HT and MET bins, if there are 4 HT and 4 MET bins
(including eff and btag corrections).

  toytt->Draw("(fit_susy_0lep_3da[2][2][2]+fit_susy_0lep_3da[2][3][2]+fit_susy_0lep_3da[3][2][2]+fit_susy_0lep_3da[3][3][2])")


Fit value of the total number of SUSY 0lep events in
the highest two HT and MET bins, if there are 4 HT and 4 MET bins
(including eff and btag corrections) integrating over the
nbjet multiplicity.

  toytt->Draw("Sum$(fit_susy_0lep_3da[2][2][]+fit_susy_0lep_3da[2][3][]+fit_susy_0lep_3da[3][2][]+fit_susy_0lep_3da[3][3][])")